Move reports to their own file
